move PageRedirectException to internal.exception package
