+++ components/uuf-core/src/main/java/org/wso2/carbon/uuf/internal/RequestDispatcher.java
@@ -24,13 +24,13 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.wso2.carbon.uuf.api.config.Configuration;
 import org.wso2.carbon.uuf.core.App;
-import org.wso2.carbon.uuf.exception.PageRedirectException;
 import org.wso2.carbon.uuf.exception.UUFException;
 import org.wso2.carbon.uuf.internal.debug.DebugLogger;
 import org.wso2.carbon.uuf.internal.debug.Debugger;
 import org.wso2.carbon.uuf.internal.deployment.AppRegistry;
 import org.wso2.carbon.uuf.internal.exception.DeploymentException;
 import org.wso2.carbon.uuf.internal.exception.HttpErrorException;
+import org.wso2.carbon.uuf.internal.exception.PageRedirectException;
 import org.wso2.carbon.uuf.internal.filter.CsrfFilter;
 import org.wso2.carbon.uuf.internal.filter.Filter;
 import org.wso2.carbon.uuf.internal.filter.FilterResult;
